+++ server/app.py
@@ -8,7 +8,7 @@ from datetime import timedelta
 
 import sys
 from pathlib import Path
-from models import db
+from model import db
 from views import s3
 
 from views.cctvAPI import CCTVS, CCTVList
