--- server/models/__init__.py
@@ -1,7 +1,0 @@
-# ORM
-
-from flask_sqlalchemy import SQLAlchemy
-db = SQLAlchemy()
-
-from flask_marshmallow import Marshmallow
-ma = Marshmallow()
