+++ server/views/memberAPI.py
@@ -3,8 +3,8 @@ import bcrypt, jwt
 from flask import Response, jsonify, make_response, redirect, url_for, session
 from flask_restful import Resource, reqparse
 from sqlalchemy.exc import IntegrityError
-from models import db
-from models.member import Member, MemberSchema
+from model import db
+from model.member import Member, MemberSchema
 
 # DB에서 '/Members/<member_id>'로 URL쳤을때 해당 멤버 정보 가져오는 부분
 class Members(Resource):
